Moved slider enums to lib
