+++ src/lib/enums/appEnum.ts
@@ -20,6 +20,17 @@ export enum colorEnum {
   DEFAULT = 'default'
 }
 
+export enum orientationEnum {
+  HORIZONTAL = 'horizontal',
+  VERTICAL = 'vertical',
+}
+
+export enum valueLabelDisplayEnum {
+  ON = 'on',
+  AUTO = 'auto',
+  OFF = 'off'
+}
+
 // Size ENUM for switch currently.
 // Note: Switch can't use LARGE, so create new enum if more properties are needed, else typescript will error
 export enum sizeEnum {
